backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7
gspread==5.12.0
oauth2client==4.1.3
python-dotenv==1.0.0
google-generativeai==0.3.1
uuid==1.30